verity/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20
python-dotenv==1.0.1
pydantic==2.10.4
tavily-python
firecrawl-py
gradient==1.0.0
openai==1.59.4
httpx==0.27.2
aiofiles==24.1.0
pillow==11.1.0
